Refactor GQL hook to be inlined with consuming components
